Change route to include the format as well
This patch alters the exposed route to comprise the sought format as
well. The new route is /:format/:title/:revid/:id , where :id is allowed
to contain the trailing file extension as well. If it is provided, it
must match the given format. A few tests checking this as well as the
format itself have been added.

Bug: T97123
Change-Id: I093f1e0a4aad07477f1bfe0cc2b5cd8d7b9efe57