Minor: do not check body against the RegEx
Grpahoid sends out binary blobs, which often cannot be UTF-decoded, thus
producing false negatives for the monitoring utility. Furthermore,
checking binary data against a RegEx is not really a sane thing to do,
so drop the body: /.+/ check. We can still rely on the response code and
Content-Type header checks to detect anomalies.

Change-Id: I22c97261c339013309e1ae11e6be9881d7c20978